typecast int -> uint in dropdown

+++ src/ui/dropdown.c
@@ -55,7 +55,7 @@ void dropdown_drawactive(void) {
         }
     }
 
-    if (y + h * drop->dropcount > settings.window_height) {
+    if (y + h * drop->dropcount > (int)settings.window_height) {
         // y -= h * (drop->dropcount - 1);
         // sign = -1;
     }
@@ -127,7 +127,7 @@ bool dropdown_mmove(DROPDOWN *d, int UNUSED(x), int y, int w, int h, int mx, int
     if (d->open) {
         int over = my / h + d->selected;
 
-        if (y + h * d->dropcount > settings.window_height) {
+        if (y + h * d->dropcount > (int)settings.window_height) {
             // over = my > 0 ? 0 : ((-my) / h + 1);
         }
